changed project structure
picked out Token class, TokenType and MdPosition enums to different files and put them in one folder
