+++ cs/Markdown/Markdown/Markdown.csproj
@@ -56,10 +56,13 @@
   <ItemGroup>
     <Compile Include="Md.cs" />
     <Compile Include="MdElement.cs" />
+    <Compile Include="Tokenizer\MdPosition.cs" />
     <Compile Include="Tests\MdTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Tests\TokenizerTests.cs" />
-    <Compile Include="Tokenizer.cs" />
+    <Compile Include="Tokenizer\Token.cs" />
+    <Compile Include="Tokenizer\Tokenizer.cs" />
+    <Compile Include="Tokenizer\TokenType.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
