+++ cs/Markdown/Markdown/Tokenizer/TokenType.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Markdown
+{
+    public enum TokenType
+    {
+        Text,
+        MdElement,
+        WhiteSpace
+    }
+}
